requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
colorama==0.4.4
dblpy==0.4.0
discord==1.0.1
discord.py==1.5.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
mysql-connector==2.2.9
pylint==2.6.0
python-dotenv==0.15.0
ratelimiter==1.2.0.post0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
yarl==1.5.1